+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_start/setup.js
@@ -350,11 +350,11 @@ if (!gwoSetupLoaded) {
                 // Minions don't have a faction number so use the previous one
                 // which should be from the primary AI and accurate
                 aiFaction = ai.faction ? ai.faction : aiFaction;
-                var legonisMachinaTags = ["tank", "lateorbital"];
-                var foundationTags = ["air", "lateorbital"];
-                var synchronousTags = ["bot", "lateorbital"];
-                // revenants are differentiated through not using "lateorbital"
-                var clusterTags = ["land", "lateorbital"];
+                var legonisMachinaTags = ["tank"];
+                var foundationTags = ["air"];
+                var synchronousTags = ["bot"];
+                var revenantsTags = ["orbital"];
+                var clusterTags = ["land"];
                 switch (aiFaction) {
                   case 0:
                     ai.personality.personality_tags =
@@ -369,6 +369,10 @@ if (!gwoSetupLoaded) {
                   case 2:
                     ai.personality.personality_tags =
                       ai.personality.personality_tags.concat(synchronousTags);
+                    break;
+                  case 3:
+                    ai.personality.personality_tags =
+                      ai.personality.personality_tags.concat(revenantsTags);
                     break;
                   case 4:
                     ai.personality.personality_tags =